requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Werkzeug==2.2.2
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy